Provide mechanism for skipping retries w/o acknowledging job
Prior to this commit a job could fail in one of two ways: it could
fail with an Unrecoverable error, in which case it'd be ack'd and
removed from the queue; or it could fail with another error and be
retried with exponential backoff.

There are some situations where neither is desirable; for example, you
might want to delay processing a job by manually re-Enqueuing it
without exponential backoff.

This commit introduces a new error, ErrDoNotRetry, and makes
ErrUnrecoverable wrap it. Rather than testing for equality to
ErrUnrecoverable when determining if we should retry, we check if the
error Is (or wraps) ErrDoNotRetry.

This allows the handler (or more likely handler middleware) to more
precisely control the processing of a job.